requirements.txt
numpy>=1.25.2
Pillow>=10.0.0
pytest==7.4.1
pytest-cov==4.1.0
